% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alert_columns.R
\name{add_alert_columns}
\alias{add_alert_columns}
\title{Title add_alert_columns}
\usage{
add_alert_columns(basedf, alertdf)
}
\value{
dataframe
}
\description{
Attach alert unique identifier columns from the alert dataset matched by alert_id
}
