% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alert_stringency.R
\name{add_alert_stringency}
\alias{add_alert_stringency}
\title{Title add_alert_stringency}
\usage{
add_alert_stringency(basedf)
}
\arguments{
\item{basedf}{}
}
\value{
dataframe
}
\description{
Attach numeric alert stringency value column (low = 1, high = 7), based on numeric alert_number
}
