% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outcome_bin.R
\name{add_outcome_bin}
\alias{add_outcome_bin}
\title{Title add_outcome_bin}
\usage{
add_outcome_bin(basedf, bins)
}
\arguments{
\item{basedf}{}

\item{bins}{}
}
\value{
dataframe
}
\description{
bin case outcomes
}
