% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pop3_adm_columns_outbreak.R
\name{add_pop3_adm_columns_outbreak}
\alias{add_pop3_adm_columns_outbreak}
\title{Title add_pop3_adm_columns_outbreak}
\usage{
add_pop3_adm_columns_outbreak(basedf, popdf)
}
\arguments{
\item{basedf}{a df with a location and outbreak_start column and existing population column}
}
\value{
the df with an alert_id column with added columns for population size and spatial scale
}
\description{
Attach population and spatial scale columns from pre-outbreak extraction to a dataset with an alert_id
}
