% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pop6_adm_columns.R
\name{add_pop6_adm_columns}
\alias{add_pop6_adm_columns}
\title{Title add_pop6_adm_columns}
\usage{
add_pop6_adm_columns(basedf, popdf)
}
\arguments{
\item{basedf}{a df with alert_id column, can be an alert or an alert group df}

\item{popdf}{a pre-outbreak df}
}
\value{
the df with an alert_id column with added columns for population size and spatial scale
}
\description{
Attach population and spatial scale columns from pre-outbreak extraction to a dataset with an alert_id
}
