% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_duplicate_observations.R
\name{average_duplicate_observations}
\alias{average_duplicate_observations}
\title{average_duplicate_observations}
\usage{
average_duplicate_observations(original_data, ...)
}
\arguments{
\item{original_data}{dataframe with location, epiweek, TL, TR, sCh, cCh, deaths}
}
\description{
when observations with the same location and overlapping epiweeks exist, average them and remove the duplicates.
}
