% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cases.R
\name{calculate_cases}
\alias{calculate_cases}
\title{Title calculate_cases}
\usage{
calculate_cases(
  alert_group_df,
  preoutbreak_ts,
  delay_period,
  evaluation_duration
)
}
\arguments{
\item{alert_group_df}{df with alert groups identified by their alert_id (one per row), location, date of first alert, date of last alert, alert number, alert type}

\item{preoutbreak_ts}{preoutbreak time series}

\item{evaluation_duration}{the duration (weeks) of the evaluation period}

\item{delay}{the delay (weeks) after the start of the alert group that will begin the evaluation period}
}
\value{
the alert group df with the alert_id, location, evaluation period start, evaluation period end, and total number of suspected cases during the evaluation period
}
\description{
Calculates the number of cases in the evaluation period for each alert group
}
