% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_population_density.R
\name{calculate_population_density}
\alias{calculate_population_density}
\title{Title calculate_population_density}
\usage{
calculate_population_density(time_series, outbreak_shp)
}
\arguments{
\item{time_series}{time series data frame for each location/week inc. location period column}

\item{outbreak_shp}{sf dataframe with shapefiles per location period}
}
\value{
time series data frame with added area per 1km^2 and population density per 1km^2 columns
}
\description{
Title calculate_population_density
}
