% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_location_names.R
\name{clean_location_names}
\alias{clean_location_names}
\title{clean_location_names}
\usage{
clean_location_names(original_data)
}
\description{
this function is used to clean the locations pulled from taxonomy database (used in clean_psql_data)
}
