% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_psql_data.R
\name{clean_psql_data}
\alias{clean_psql_data}
\title{clean_psql_data}
\usage{
clean_psql_data(original_data, ...)
}
\arguments{
\item{original_data}{original dataset}
}
\description{
this function is used to clean the raw data pulled from psql database
}
