% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alert_groups2.R
\name{create_alert_groups2}
\alias{create_alert_groups2}
\title{Title create_alert_groups2}
\usage{
create_alert_groups2(alerts, nweeks)
}
\arguments{
\item{alerts}{a df with formatted alerts}

\item{nweeks}{the number of weeks where an alert of a specific number has not been triggered so that this alert is considered postprocessed (starts a new alert group)}
}
\value{
the alert group df, each row is a unique alert group. Includes the alert_id of the first alert in the group, the TL of the first and last alert in the group, the alert number and alert type
}
\description{
Title create_alert_groups2
}
