% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_aggregator.R
\name{custom_paste}
\alias{custom_paste}
\title{custom_paste}
\usage{
custom_paste(x)
}
\description{
This is an helper function to make sure observation_collection_id columns are pasted in order and only include a set of unique OC UIDs.
}
