% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_postprocessed_alerts.R
\name{define_postprocessed_alerts}
\alias{define_postprocessed_alerts}
\title{Title define_postprocessed_alerts}
\usage{
define_postprocessed_alerts(alerts_df, nweeks)
}
\arguments{
\item{alerts_df}{each row represents a unique alert}

\item{nweeks}{the number of weeks where an alert of a specific number has not been triggered so that this alert is considered postprocessed (starts a new alert group)}
}
\value{
alerts df with an added column (logical) indicating whether the alert is postprocessed (marks the start of a new alert group)
The new column is used to create alert groups with OutbreakExtractR::create_alert_groups2
}
\description{
Title define_postprocessed_alerts
}
