% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop.R
\name{estimate_adj_factors}
\alias{estimate_adj_factors}
\title{Estimate Adjustment Factors for Population Data (for years >=2021, use the adjustment factors at 2020 instead)}
\usage{
estimate_adj_factors(
  country,
  year,
  country_shp = NULL,
  raster_dir = "data/raster",
  dest_dir = "data"
)
}
\arguments{
\item{country}{character: countries whose population needs to be estimated.}

\item{year}{numeric vector: Years for which population needs to be estimated.}

\item{country_shp}{sf object: shapefiles of the country}

\item{raster_dir}{string: Directory containing the WorldPop raster files.}

\item{dest_dir}{string: Directory to download missing raster files.}
}
\value{
population adjustment factor
}
\description{
Estimate Adjustment Factors for Population Data (for years >=2021, use the adjustment factors at 2020 instead)
}
