% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_agroup_case_outcomes_obs.R
\name{extract_agroup_case_outcomes}
\alias{extract_agroup_case_outcomes}
\title{extract_agroup_case_outcomes}
\usage{
extract_agroup_case_outcomes(
  dat,
  alert_col,
  alert_group_col,
  delay_days,
  evalperiod_days,
  intervene_thresh,
  out_col
)
}
\arguments{
\item{dat}{dataframe}

\item{alert_col}{character string of column name that contains unique alert id}

\item{alert_group_col}{character string of column name that contains alert group id}

\item{delay_days}{start evaluation period <delay_days> after alert is triggered}

\item{evalperiod_days}{evaluation period duration in days}

\item{intervene_thresh}{number of sCh above which it's useful to intervene with a campaign}

\item{out_col}{outcome ID}
}
\description{
calculate case-related outcomes for alert groups, incl sCh, cCh, deaths, and count number of weeks with non-zero sCh for each location, assuming <delay_days> delay to OCV implementation and <evalperiod_days> eval period.
}
