% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_agroup_pers_outcomes_obs.R
\name{extract_agroup_pers_outcomes}
\alias{extract_agroup_pers_outcomes}
\title{extract_agroup_pers_outcomes}
\usage{
extract_agroup_pers_outcomes(
  dat,
  alert_col,
  alert_group_col,
  delay_days,
  evalperiod_days,
  nweek_thresh,
  out_settings
)
}
\arguments{
\item{dat}{dataframe}

\item{alert_col}{character string of column name that contains unique alert id}

\item{alert_group_col}{character string of column name that contains alert group id}

\item{delay_days}{start evaluation period <delay_days> after alert is triggered}

\item{evalperiod_days}{evaluation period duration in days}

\item{nweek_thresh}{number of weeks with X sCh cases in the next year above which it's useful to intervene with a campaign}

\item{out_col}{outcome ID}
}
\description{
calculate persistence-related outcomes for alert groups, incl nweeks_overX_sCh (encoded in metric & value), total_weeks in the year TS, assuming <delay_days> delay to OCV implementation and <evalperiod_days> eval period.
}
