% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_binary_case_outcomes.R
\name{extract_binary_case_outcomes}
\alias{extract_binary_case_outcomes}
\title{Title extract_binary_case_outcomes}
\usage{
extract_binary_case_outcomes(eval_case_df, case_thresh = 200)
}
\arguments{
\item{eval_case_df}{a df with the number of cases in the evaluation period per alert group}

\item{case_thresh}{case threshold that determines an outcome as TRUE/FALSE}
}
\value{
dataframe
}
\description{
returns a df with a binary outcome per alert group that is TRUE if the number of cases in the evaluation period has exceeded the threshold
}
