% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_lps.R
\name{fill_missing_lps}
\alias{fill_missing_lps}
\title{fill_missing_lps}
\usage{
fill_missing_lps(original_data)
}
\arguments{
\item{original_data}{dataframe with location, epiweek, TL, TR, sCh, cCh, deaths}
}
\description{
If an observation is missing a location period ID but there is another observation in the dataset with a location period ID for that location, fill in the missing location period ID. If there are multiple lp ID options, select the lp ID from the most recent observation.
}
