% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_phantom_zeroes.R
\name{fill_phantom_zeroes}
\alias{fill_phantom_zeroes}
\title{fill_phantom_zeroes}
\usage{
fill_phantom_zeroes(original_data)
}
\arguments{
\item{original_data}{dataframe with location, epiweek, TL, TR, sCh, cCh, deaths}
}
\description{
Fill in weekly zeroes for sCh if no reporting between min(TL)- 8 weeks and max(TL)+ 8 weeks for that location
}
