% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_alerts.R
\name{format_alerts}
\alias{format_alerts}
\title{Title format_alerts}
\usage{
format_alerts(alerts_df)
}
\value{
a 'long' version of the alerts dataframe with columns for the alert_id, location, TL, spatial scale, country, alert number (numeric), and alert type
}
\description{
Creates a 'long' version of the alerts df (one row per alert) and removes redundant information (this info can be found in the preoutbreak time series file)
}
