% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epiweek.R
\name{get_epiweek}
\alias{get_epiweek}
\title{get_epiweek}
\usage{
get_epiweek(epidate)
}
\arguments{
\item{epidate}{date object to}
}
\value{
epiweek string (YYYY-ww) where "ww" means 2-digit week number
}
\description{
Return epiweek column for a given date
}
