% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghs_pop.R
\name{get_ghs_pop}
\alias{get_ghs_pop}
\title{get_ghs_pop}
\usage{
get_ghs_pop(shp, years, raster_dir, crs = 4326)
}
\arguments{
\item{shp}{sf object: geometries.}

\item{years}{numeric vector: Years for which population needs to be estimated.}

\item{raster_dir}{string: Directory containing the downloaded GHS raster files.}

\item{crs}{numeric: Coordinate reference system for the geometry (default: EPSG 4326).}
}
\value{
A dataframe with target year and estimated population.
}
\description{
Estimate yearly population based on geometry and GHS population rasters.
}
