% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_threshold.R
\name{get_outbreak_threshold}
\alias{get_outbreak_threshold}
\title{get_outbreak_threshold}
\usage{
get_outbreak_threshold(
  threshold_type = c("fixed threshold", "mean weekly incidence rate",
    "outbreak_dependent threshold", "time_restricted threshold"),
  fixed_outbreak_threshold = NULL,
  surveillance_data = NULL,
  zero_case_assumption = FALSE,
  customized_TL = NULL,
  customized_TR = NULL
)
}
\arguments{
\item{surveillance_data}{dataframe: original cholera surveillance data}

\item{threahold_type:}{character: 1. fixed threshold: a fixed value as outbreak threshold. 2. mean weekly incidence: use the mean weekly incidence as the outbreak threshold. 3. outbreak_dependent threshold: use the mean weekly cholera incidence for the first three weeks as the threshold for that outbreak}

\item{fixed_outbreak_threshold:}{numeric: the assigned outbreak threshold (cases per week)}

\item{zero_case_assumption:}{logic: whether to assume weeks without reports have zero case}

\item{customized_TL:}{customize the lower bound of time for outbreak estimation}

\item{customized_TR:}{customize the upper bound of time for outbreak estimation}
}
\value{
dataframe
}
\description{
this function will be used in the identify_outbreak** functions. 1. average weekly/daily incidence across the whole study time period. 2. a fixed number. 3. changing outbreak threshold (for instance: the average weekly/daily cases of the first three weeks).
}
