% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniform_wday_startdate.R
\name{get_uniform_wday_startdate}
\alias{get_uniform_wday_startdate}
\title{get_uniform_wday_startdate}
\usage{
get_uniform_wday_startdate(x, wstart)
}
\arguments{
\item{x}{date}

\item{wstart}{number representing the day of the week that should be considered the week start (1 = Monday)}
}
\description{
Returns the week's starting date for the date that is provided
}
