% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_consecutive_outbreak_data.R
\name{identify_consecutive_outbreak_data}
\alias{identify_consecutive_outbreak_data}
\title{identify_consecutive_outbreak_data}
\usage{
identify_consecutive_outbreak_data(
  outbreak_data,
  minimum_consecutive_reports = 3,
  temporal_scale = c("weekly"),
  cons_group_col = "consecutive_group",
  cons_obs_col = "consecutive_obs"
)
}
\description{
this function will be used in identify_epidemic_start and identify_epidemic_period functions
}
