% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_epidemic_period.R
\name{identify_epidemic_period}
\alias{identify_epidemic_period}
\title{identify_epidemic_period}
\usage{
identify_epidemic_period(
  outbreak_data = outbreak_with_start,
  epidemic_tail = 6,
  max_missingdata_period = 12
)
}
\description{
this function will be used in identify_outbreak** functions. This function will separate different outbreaks in the time series data for a certain location with a certain threshold.
}
