% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_epidemic_tail.R
\name{identify_epidemic_tail}
\alias{identify_epidemic_tail}
\title{identify_epidemic_tail}
\usage{
identify_epidemic_tail(outbreak_data, tail_period = 6)
}
\description{
this function will be used in identify_outbreak** functions
}
