% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_washout_period.R
\name{identify_washout_period}
\alias{identify_washout_period}
\title{identify_washout_period}
\usage{
identify_washout_period(outbreak_data, washout_period = 3, tail_period = 2)
}
\description{
this function will be used in identify_outbreak** functions
}
