% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_weeks_exceeding_threshold.R
\name{identify_weeks_exceeding_threshold}
\alias{identify_weeks_exceeding_threshold}
\title{identify_weeks_exceeding_threshold}
\usage{
identify_weeks_exceeding_threshold(
  outbreak_data,
  threshold_col = c("case_threshold"),
  threshold_type = c("case", "rate"),
  pop_col = NULL
)
}
\description{
identify_weeks_exceeding_threshold
}
