% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_alerts_short.R
\name{label_alerts_short}
\alias{label_alerts_short}
\title{Title label_alerts_short}
\usage{
label_alerts_short(basedf, is_ordered = FALSE)
}
\arguments{
\item{basedf}{a df with a column for alert numbers}

\item{is_ordered}{logical with default value of FALSE}
}
\value{
dataframe
}
\description{
function that adds short descriptive labels to alert numbers
}
