% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_alerttypes.R
\name{label_alerttypes}
\alias{label_alerttypes}
\title{Title label_alerttypes}
\usage{
label_alerttypes(basedf)
}
\arguments{
\item{basedf}{a df with a column for alert_type}
}
\value{
dataframe
}
\description{
function that converts alert types to an ordered factor with pretty labels
}
