% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_dimensions.R
\name{label_dimensions}
\alias{label_dimensions}
\title{Title label_dimensions}
\usage{
label_dimensions(basedf, is_ordered = TRUE)
}
\arguments{
\item{basedf}{a df with a column for dimension}

\item{is_ordered}{logical with default value of TRUE}
}
\value{
dataframe
}
\description{
function that adds figure-worthy utility dimension labels
}
