% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_aggregator.R
\name{observation_aggregator}
\alias{observation_aggregator}
\title{observation_aggregator}
\usage{
observation_aggregator(daily_outbreak_data)
}
\description{
aggregate daily cholera data to weekly cholera data
}
