% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_filter.R
\name{observation_filter}
\alias{observation_filter}
\title{observation_filter}
\usage{
observation_filter(
  outbreak_data,
  time_lower_bound_filter = lubridate::ymd("2000-01-01"),
  time_upper_bound_filter = lubridate::ymd("2023-12-31"),
  temporal_scale_filter = c("daily", "weekly"),
  who_regions = c("AFR", "EMR", "AMR"),
  spatial_scale_filter = c("country", "admin1", "admin2", "admin3"),
  remove_na_sCh = TRUE,
  remove_na_cCh = FALSE,
  remove_na_locationperiod = FALSE,
  minimum_daily_cases = 0,
  ...
)
}
\description{
filter observations base on certain criteria
}
