% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_consecutive_alerts_obs.R
\name{remove_consecutive_alerts}
\alias{remove_consecutive_alerts}
\title{remove_consecutive_alerts}
\usage{
remove_consecutive_alerts(alerts_df, nweeks = 4)
}
\arguments{
\item{alerts_df}{dataframe with alerts}

\item{nweeks}{number of weeks of consecutive FALSE required for an alert to be TRUE}
}
\value{
dataframe with columns with the pre-processed and post-processed alerts
}
\description{
Removes consecutive alerts of the same type for alerts 0-17. An alert can only be TRUE if it is preceded by nweeks of consecutive FALSE, which ensures there are no consecutive weeks of the same alert type.
}
