% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_alert.R
\name{trigger_alert}
\alias{trigger_alert}
\title{trigger_alert}
\usage{
trigger_alert(original_data)
}
\arguments{
\item{original_data}{dataframe of pre-alert time series extractions}
}
\value{
dataframe
}
\description{
Triggers alerts according to patterns in suspected case trends. Returns a dataframe with alert indicators aligned with the pre-alert timeseries extractions across multiple locations. Alert is TRUE for the week when the alert would have been identified. Multiple alerts are currently identified, some of which were transferred from trigger_alert_caseratio and trigger_alert_cases. Alerts are triggered in the week AFTER the following patterns are identified: 1-3 TREND-BASED)  number of cases in 1/2/3 weeks in the past month exceeded the past 4-weeks rolling case mean, thus indicating an increase in cases relative to the past month; 4-10 CASE-BASED) at least 2/5/10/25/50/100/250 weekly suspected cases in 3 consecutive weeks; 11-17 CUM CASE-BASED) at least 5/10/25/50/100/500/1000 cumulative suspected cases in past 3 weeks.
}
