% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_alert_cases_obs.R
\name{trigger_alert_cases}
\alias{trigger_alert_cases}
\title{trigger_alert_cases}
\usage{
trigger_alert_cases(original_data)
}
\arguments{
\item{original_data}{dataframe of pre-alert time series extractions}
}
\value{
dataframe
}
\description{
Triggers alerts according to patterns in suspected cases. Returns a dataframe with alert indicators aligned with the pre-alert timeseries extractions across multiple locations. Alert is TRUE for the week when the alert would have been identified. Three alerts are currently identified: 1) four 0s followed by 2 consecutive weeks with at least 2 sCh each; 2) four 0s followed by a week with at least 2 sCh and at least 10 sCh in the 5-week period but one of the weeks reported 0 cases; 3) <4 sCh in 4 consecutive weeks followed by 2 consecutive weeks with at least 2 sCh each
}
