% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_alert_rate.R
\name{trigger_alert_rate}
\alias{trigger_alert_rate}
\title{trigger_alert_rate}
\usage{
trigger_alert_rate(original_data)
}
\arguments{
\item{original_data}{dataframe of pre-alert time series extractions}
}
\value{
dataframe
}
\description{
Triggers rate-based alerts 18-24 according to patterns in incidence rate trends per 10000 people
}
