% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_rate_alert.R
\name{trigger_rate_alert}
\alias{trigger_rate_alert}
\title{trigger_rate_alert}
\usage{
trigger_rate_alert(original_data)
}
\arguments{
\item{original_data}{dataframe of pre-alert time series extractions}
}
\value{
dataframe
}
\description{
Triggers rate-based alerts 18-24 according to patterns in incidence rate trends per 10000 people
}
