% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alert_columns_outbreak.R
\name{add_alert_columns_outbreak}
\alias{add_alert_columns_outbreak}
\title{Title add_alert_columns_outbreak}
\usage{
add_alert_columns_outbreak(basedf, alertdf)
}
\value{
dataframe
}
\description{
Attach alert unique identifier columns to the linked / outbreak datasets
}
