% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_country_column.R
\name{add_country_column}
\alias{add_country_column}
\title{Title add_country_column}
\usage{
add_country_column(basedf)
}
\arguments{
\item{basedf}{}
}
\value{
dataframe
}
\description{
Attach column with ISO3 country code (country) which is derived from the basedf location
}
