% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_negative_alert_groups.R
\name{add_negative_alert_groups}
\alias{add_negative_alert_groups}
\title{Title add_negative_alert_groups}
\usage{
add_negative_alert_groups(summary_outcomes)
}
\arguments{
\item{summary_outcomes}{A dataframe containing alert group information and outcomes}
}
\value{
dataframe with summary outcomes for both positive and added rows for negative alert groups. Negative alert groups get their outcome and out_type from the corresponding alert group type 0 that occurred in the same location and epiweek
}
\description{
Takes a dataframe with summary outcomes for positive alerts and alert groups (created with extract_agroup_case_outcomes) and creates negative alert groups for alert groups 1-5.
Negative alerts groups occur if an alert group 1-5 was not trigerred while an alert group 0 was at a specific location-epiweek combination.
Negative alerts groups inherit the outcome and outcome_type from the corresponding alert 0. If the outcome is TRUE, they are considered false negatives.
If the outcome is FALSE, they are considered true negatives.
}
