% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unique_alert_ids.R
\name{add_unique_alert_ids}
\alias{add_unique_alert_ids}
\title{add_unique_alert_ids}
\usage{
add_unique_alert_ids(dat)
}
\arguments{
\item{dat}{dataframe}
}
\description{
add unique IDs for each alert -- assuming alerts 0-17 based on trigger_alert function
}
