% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alert_groups.R
\name{create_alert_groups}
\alias{create_alert_groups}
\title{create_alert_groups}
\usage{
create_alert_groups(df, alert_cols)
}
\arguments{
\item{df}{dataframe with alerts (has columns with pre-processed and post-processed alerts)}

\item{alert_cols}{a vector with the names of the alert columns}
}
\value{
a dataframe with the existing columns of df plus an alert group id column. The alert group id includes the location and date (epiweek) of the first alert in the alert group
}
\description{
Group alerts into alert groups that include only one postprocessed alert of the same type and 0 or more preprocessed alerts
The groups are asigned an alert_group_id
}
