% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop.R
\name{get_pop}
\alias{get_pop}
\title{get_pop}
\usage{
get_pop(
  shp = shp,
  year = 2000,
  country = "AGO",
  country_shp = NULL,
  pop_raster_path = "worldpop",
  raster_dir = NULL
)
}
\arguments{
\item{shp}{sf objects: geometries}

\item{year}{numeric vector: Years for which population needs to be estimated.}

\item{country}{character: countries whose population needs to be estimated.}

\item{country_shp}{sf object: country geometry}

\item{raster_dir}{path to save the GHS pop raster file for a certain year}

\item{pop_raster_path:}{path to save the pop raster file for a certain year (Estimated total number of people per grid-cell)}
}
\value{
numeric
}
\description{
this function is used to get population data for each location period
}
