% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shp.R
\name{get_shp}
\alias{get_shp}
\title{get_shp}
\usage{
get_shp(
  location_period_id,
  username = username,
  password = password,
  host = "db.cholera-taxonomy.middle-distance.com",
  port = 5432,
  dbname = "CholeraTaxonomy_production"
)
}
\description{
this function is to extract shapefiles from
}
