% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_epidemic_start.R
\name{identify_epidemic_start}
\alias{identify_epidemic_start}
\title{identify_epidemic_start}
\usage{
identify_epidemic_start(
  outbreak_data = outbreak_data,
  minimum_consecutive_reports = 3
)
}
\description{
this function will be used in identify_outbreak** functions. This function is mainly to check the consecutively increasing epiweeks.

per location period per start_day group: weekly cholera incidence
}
