% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_outbreaks.R
\name{identify_outbreaks}
\alias{identify_outbreaks}
\title{identify_outbreaks}
\usage{
identify_outbreaks(
  threshold_type,
  original_data,
  zero_case_assumption = T,
  customized_TL = NULL,
  customized_TR = NULL
)
}
\arguments{
\item{original_data}{dataframe: original cholera surveillance data}

\item{threahold_type:}{character: 1. fixed threshold: a fixed value as outbreak threshold. 2. mean weekly incidence: use the mean weekly incidence as the outbreak threshold. 3. outbreak_dependent threshold: use the mean weekly cholera incidence for the first three weeks as the threshold for that outbreak}

\item{fixed_outbreak_threshold:}{numeric: the assigned outbreak threshold (cases per week)}

\item{zero_case_assumption:}{logic: whether to assume weeks without reports have zero case}

\item{customized_TL:}{customize the lower bound of time for outbreak estimation}

\item{customized_TR:}{customize the upper bound of time for outbreak estimation}
}
\value{
list
}
\description{
identify_outbreaks
}
