% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_alert_caseratio.R
\name{trigger_alert_caseratio}
\alias{trigger_alert_caseratio}
\title{trigger_alert_caseratio}
\usage{
trigger_alert_caseratio(original_data)
}
\arguments{
\item{original_data}{dataframe of pre-alert time series extractions}
}
\value{
dataframe
}
\description{
Triggers alerts according to patterns in excess suspected cases relative to a historical 4-week case threshold. Returns a dataframe with alert indicators aligned with the pre-alert timeseries extractions across multiple locations. Alert is TRUE for the week when the alert would have been identified. Three alerts are currently identified: 1) alert week always exceeds mean of last 4 weeks of cases, thus indicating an increase in cases relative to the past month; 2) exceeds mean of last 4 weeks of cases in alert week and at least one other time within the past month; 3) exceeds mean of last 4 weeks of cases in alert week and at least two other times within the past month
}
